Pose Estimation Dashboard/requirements.txt
streamlit
opencv-python
mediapipe
plotly
pillow
pandas
numpy

